requirements.txt

### 2. requirements.txt (依赖清单)
langchain-mcp-adapters>=0.1.0
langgraph>=0.2.0
langchain-openai>=0.1.0
fastmcp>=0.1.0
aiohttp>=3.9.0
python-dotenv>=1.0.0